requirements.txt
aiohttp
aiohttp-jinja2
jinja2
pyserial